fix: update smart contract
